requirements.txt
# requirements.txt para prueba.py (FastAPI + OCR + PDF + SRI, CPU only)

# Web API
fastapi==0.116.1
uvicorn==0.35.0

# PDF parsing
pdfminer.six==20250506
PyMuPDF==1.26.4
pdfplumber==0.11.7
Pillow==11.3.0
pikepdf==9.4.2

# OCR
easyocr==1.7.2
pytesseract==0.3.13
# Torch CPU-only (más liviano)
torch==2.4.0+cpu
torchvision==0.19.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
numpy==2.2.6
opencv-python-headless==4.12.0.88
python-bidi==0.6.6

# Barcode/QR code reading
pyzbar==0.1.9

# Imagen hashing
imagehash==4.3.1     # <--- 🚨 agregado

exifread==3.0.0     # <--- 🚨 agregado

piexif==1.1.3     # <--- 🚨 agregado

python-multipart==0.0.9 

# SRI SOAP client
zeep==4.3.1
requests==2.32.5
lxml==6.0.1
isodate==0.7.2
requests-toolbelt==1.0.0
requests-file==2.1.0

# AWS SDK
boto3==1.34.162
botocore==1.34.162

# Utilidades varias
pydantic==2.11.7
typing_extensions==4.15.0
python-dateutil==2.9.0.post0